correct prototype to use size_t type

+++ bld/wgml/c/wgmlmsg.c
@@ -69,9 +69,9 @@ bool init_msgs( void )
 /*  get a msg text string                                                  */
 /***************************************************************************/
 
-bool get_msg( msg_ids resid, char *buff, int buff_len )
+bool get_msg( msg_ids resid, char *buff, size_t buff_len )
 {
-    if( hInstance.status == 0 || WResLoadString( &hInstance, resid + MsgShift, buff, buff_len ) <= 0 ) {
+    if( hInstance.status == 0 || WResLoadString( &hInstance, resid + MsgShift, buff, (int)buff_len ) <= 0 ) {
         buff[0] = '\0';
         return( false );
     }
